Renamed custom_user folder ==> custom_models + Renamed models ==> user_models + Reseted all the migrations and migrated them back with the changes - issues: database check model is not recognized
