+++ backend/custom_models/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import User
+from .user_models import User
 from django.contrib.auth.admin import UserAdmin
 
 admin.site.register(User, UserAdmin)
