--- backend/custom_user/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CustomUserConfig(AppConfig):
-    name = 'custom_user'
